+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - Add a way to follow container logs, the logs are now paginated
 - Make container and image ids open their respective page with details
 - Add a button to clear images build cache
+- Add a button to tag an image